[utils] Fix error when creating labels
The GrimoireUtils creation of labels does now have the same behaviour
as the general GrimoireLib's for dates. The initial date in the analysis
should be greater or equal to the proposed initial date and the last date
of the analysis should be strictly less than the last date proposed

Thus, a day is subsctracted from the last proposed date of analysis.