[scm] Add support for type of file action
In the following, it is possible to filter the dataset by the
type of file. This is done according to the type of files detected by
CVSAnalY: unknown, devel-doc, build, code, i18n, documentation, image, ui
and package.

Thus, this query is as smart as CVSAnalY detecting those types of files.