[mls] Fix error when calculating top email senders by company
This commit fixes all of the tops for the metric EmailsSenders.
The rest of the _get_top_xxx methods are removed.

There is still code to be removed in the future as passing as parameter
a new set of filters. This is not part of the current API. Howwever this
is left for compatibility purposes.