[analysis] Add the most active changesets waiting for a reviewer
This study provides the list of Gerrit changesets with the highest
activity that are still waiting for a reviewer action. Thus they
were not marked with a Code-Review of -1 or -2.

This provides the related repository, the patch title, the date
of the last upload. This is also sorted by the total number of
patchsets.